refactor code | move helpers to userCollection | trial error state update / speed
